Fix a previous merge disaster
Causing all nginx options to disappear.

+++ RPi_Cam_Web_Interface_Installer.sh
@@ -62,10 +62,11 @@ case "$1" in
         sudo apt-get purge -y nginx php5 php5-fpm php5-common php-apc gpac motion
         sudo apt-get autoremove -y
 
-        sudo rm -r /var/www/*
-        sudo rm /usr/local/bin/raspimjpeg
+        sudo rm -r /var/www/$rpicamdir/*
+        sudo rm /etc/sudoers.d/RPI_Cam_Web_Interface
+        sudo rm /usr/bin/raspimjpeg
         sudo rm /etc/raspimjpeg
-        sudo cp -r etc/rc_local_std/rc.local /etc/
+        sudo cp -r /etc/rc.local.bak /etc/rc.local
         sudo chmod 755 /etc/rc.local
 
         echo "Removed everything"
@@ -102,6 +103,7 @@ case "$1" in
         if [ ! -e /var/www/$rpicamdir/cam.jpg ]; then
           sudo ln -sf /run/shm/mjpeg/cam.jpg /var/www/$rpicamdir/cam.jpg
         fi
+
         if [ "$rpicamdir" == "" ]; then
           cat etc/apache2/sites-available/default.1 > etc/apache2/sites-available/default
         else
@@ -120,7 +122,6 @@ case "$1" in
         if [ ! -e /usr/bin/raspimjpeg ]; then
           sudo ln -s /opt/vc/bin/raspimjpeg /usr/bin/raspimjpeg
         fi
-
 
         if [ "$rpicamdir" == "" ]; then
           cat etc/raspimjpeg/raspimjpeg.1 > etc/raspimjpeg/raspimjpeg
@@ -152,44 +153,79 @@ case "$1" in
         ;;
 
   install_nginx)
-        # Update and ensure the program is not running and all prerequisites are installed
         sudo killall raspimjpeg
         git pull origin master
         sudo apt-get install -y nginx php5-fpm php5-common php-apc
 
-        # Move web interface code into place
-        sudo cp -r www/* /var/www/
-        sudo mkdir -p /var/www/media
-        sudo chown -R www-data:www-data /var/www
-        sudo mknod /var/www/FIFO p
-        sudo chmod 666 /var/www/FIFO
-        # Install nginx server file
+        sudo mkdir -p /var/www/$rpicamdir/media
+        sudo cp -r www/* /var/www/$rpicamdir/
+        if [ -e /var/www/$rpicamdir/index.html ]; then
+          sudo rm /var/www/$rpicamdir/index.html
+        fi
+        sudo chown -R www-data:www-data /var/www/$rpicamdir
+        if [ ! -e /var/www/$rpicamdir/FIFO ]; then
+          sudo mknod /var/www/$rpicamdir/FIFO p
+        fi
+        sudo chmod 666 /var/www/$rpicamdir/FIFO
+
+        if [ ! -e /var/www/$rpicamdir/cam.jpg ]; then
+          sudo ln -sf /run/shm/mjpeg/cam.jpg /var/www/$rpicamdir/cam.jpg
+        fi
+
+        if [ "$rpicamdir" == "" ]; then
+          cat etc/nginx/sites-available/picamwebint.1 > etc/nginx/sites-available/picamwebint
+        else
+          sed -e "s/www/www\/$rpicamdir/" etc/nginx/sites-available/picamwebint.1 > etc/nginx/sites-available/picamwebint
+        fi
         sudo cp -r etc/nginx/sites-available/picamwebint /etc/nginx/sites-available/picamwebint
         sudo chmod 644 /etc/nginx/sites-available/picamwebint
-        sudo ln -s /etc/nginx/sites-available/picamwebint /etc/nginx/sites-enabled/picamwebint
-        # Update nginx main config file
+        if [ ! -e /etc/nginx/sites-enabled/picamwebint ]; then
+          sudo ln -s /etc/nginx/sites-available/picamwebint /etc/nginx/sites-enabled/picamwebint
+        fi
+
         sudo sed -i "s/worker_processes 4;/worker_processes 2;/g" /etc/nginx/nginx.conf
         sudo sed -i "s/worker_connections 768;/worker_connections 128;/g" /etc/nginx/nginx.conf
         sudo sed -i "s/gzip on;/gzip off;/g" /etc/nginx/nginx.conf
         if ["$NGINX_DISABLE_LOGGING"]; then
             sudo sed -i "s:access_log /var/log/nginx/nginx/access.log;:access_log /dev/null;:g" /etc/nginx/nginx.conf
         fi
-        # Configure php-apc
+
         sudo sh -c "echo \"cgi.fix_pathinfo = 0;\" >> /etc/php5/fpm/php.ini"
         sudo cp etc/php5/apc.ini /etc/php5/conf.d/20-apc.ini
         sudo chmod 644 /etc/php5/conf.d/20-apc.ini
 
+        sudo cp etc/sudoers.d/RPI_Cam_Web_Interface /etc/sudoers.d/
+        sudo chmod 440 /etc/sudoers.d/RPI_Cam_Web_Interface
+
         sudo cp -r bin/raspimjpeg /opt/vc/bin/
         sudo chmod 755 /opt/vc/bin/raspimjpeg
-        sudo ln -s /opt/vc/bin/raspimjpeg /usr/bin/raspimjpeg
+        if [ ! -e /usr/bin/raspimjpeg ]; then
+          sudo ln -s /opt/vc/bin/raspimjpeg /usr/bin/raspimjpeg
+        fi
 
+        if [ "$rpicamdir" == "" ]; then
+          cat etc/raspimjpeg/raspimjpeg.1 > etc/raspimjpeg/raspimjpeg
+        else
+          sed -e "s/www/www\/$rpicamdir/" etc/raspimjpeg/raspimjpeg.1 > etc/raspimjpeg/raspimjpeg
+        fi
         sudo cp -r /etc/raspimjpeg /etc/raspimjpeg.bak
         sudo cp -r etc/raspimjpeg/raspimjpeg /etc/
         sudo chmod 644 /etc/raspimjpeg
 
+        if [ "$rpicamdir" == "" ]; then
+          cat etc/rc_local_run/rc.local.1 > etc/rc_local_run/rc.local
+        else
+          sed -e "s/www/www\/$rpicamdir/" etc/rc_local_run/rc.local.1 > etc/rc_local_run/rc.local
+        fi
+        sudo cp -r /etc/rc.local /etc/rc.local.bak
         sudo cp -r etc/rc_local_run/rc.local /etc/
         sudo chmod 755 /etc/rc.local
 
+        if [ "$rpicamdir" == "" ]; then
+          cat etc/motion/motion.conf.1 > etc/motion/motion.conf
+        else
+          sed -e "s/www/www\/$rpicamdir/" etc/motion/motion.conf.1 > etc/motion/motion.conf
+        fi
         sudo cp -r etc/motion/motion.conf /etc/motion/
         sudo chmod 640 /etc/motion/motion.conf
 
@@ -199,6 +235,7 @@ case "$1" in
 
         echo "Installer finished"
         ;;
+
   start)
         shopt -s nullglob
 
@@ -228,11 +265,9 @@ case "$1" in
         echo "Stopped"
         ;;
 
-
   *)
         echo "No option selected"
         ;;
 
 esac
-
